refactor: extract decodeArgs to utils
